Moved get_arg_parser to util.config
